Move Istio namer and interpreter to  own submodule
