+++ interpreter/istio/src/main/resources/META-INF/services/io.buoyant.namer.InterpreterInitializer
@@ -1,0 +1,1 @@
+io.buoyant.interpreter.k8s.istio.IstioInterpreterInitializer
