+++ interpreter/k8s/src/main/resources/META-INF/services/io.buoyant.namer.InterpreterInitializer
@@ -1,2 +1,1 @@
 io.buoyant.interpreter.k8s.ConfigMapInterpreterInitializer
-io.buoyant.interpreter.k8s.IstioInterpreterInitializer
