+++ namer/istio/src/main/resources/services/io.buoyant.namer.NamerInitializer
@@ -1,0 +1,1 @@
+io.buoyant.namer.k8s.istio.IstioInitializer
