+++ namer/istio/src/main/scala/io/buoyant/namer/k8s/istio/IstioInitializer.scala
@@ -1,9 +1,9 @@
-package io.buoyant.namer.k8s
+package io.buoyant.namer.k8s.istio
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.twitter.finagle._
 import io.buoyant.config.types.Port
-import io.buoyant.k8s.istio.{DiscoveryClient, IstioNamer}
+import io.buoyant.k8s.istio._
 import io.buoyant.namer.{NamerConfig, NamerInitializer}
 
 /**
@@ -28,7 +28,6 @@ case class IstioConfig(
   host: Option[String],
   port: Option[Port]
 ) extends NamerConfig {
-  import io.buoyant.k8s.istio._
 
   @JsonIgnore
   override val experimentalRequired = true
