+++ namer/k8s/src/main/resources/META-INF/services/io.buoyant.namer.NamerInitializer
@@ -1,4 +1,3 @@
 io.buoyant.namer.k8s.K8sExternalInitializer
 io.buoyant.namer.k8s.K8sInitializer
 io.buoyant.namer.k8s.K8sNamespacedInitializer
-io.buoyant.namer.k8s.IstioInitializer
